Save items when performing __len__ on a query
A lot of python functions (list, tuple, random.choice)
will try to do len() on an iterable before iterating/accessing it.
To make this more efficient, store the items so that we don't have
to hit the database twice.
Arguably we shouldn't support __len__ however it's a been too late
for that now...